Woo/grouped products small fixes

+++ plugins/woocommerce/src/main/kotlin/org/wordpress/android/fluxc/model/WCProductModel.kt
@@ -239,11 +239,11 @@ data class WCProductModel(@PrimaryKey @Column private var id: Int = 0) : Identif
         }
     }
 
-    fun getGroupedProductIds(): List<Int> {
-        val groupedIds = ArrayList<Int>()
+    fun getGroupedProductIds(): List<Long> {
+        val groupedIds = ArrayList<Long>()
         try {
             Gson().fromJson(groupedProductIds, JsonElement::class.java).asJsonArray.forEach { jsonElement ->
-                jsonElement.asInt.let { groupedIds.add(it) }
+                jsonElement.asLong.let { groupedIds.add(it) }
             }
         } catch (e: JsonParseException) {
             AppLog.e(T.API, e)
